% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_fit.R
\name{as.data.frame.crm_fit}
\alias{as.data.frame.crm_fit}
\title{Convert crm_fit object to \code{data.frame}.}
\usage{
\method{as.data.frame}{crm_fit}(x, ...)
}
\arguments{
\item{x}{\code{\link{crm_fit}} object to convert.}

\item{...}{Extra parameters, passed onwards.}
}
\value{
A \code{data.frame}
}
\description{
Convert crm_fit object to \code{data.frame}.
}
