% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{as.data.frame.efftox_fit}
\alias{as.data.frame.efftox_fit}
\title{Convert efftox_fit object to \code{data.frame}.}
\usage{
\method{as.data.frame}{efftox_fit}(x, ...)
}
\arguments{
\item{x}{\code{\link{efftox_fit}} object to convert.}

\item{...}{Extra parameters, passed onwards.}
}
\value{
A \code{data.frame}
}
\description{
Convert efftox_fit object to \code{data.frame}.
}
