% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_fit.R
\name{as.mcmc.list.crm_fit}
\alias{as.mcmc.list.crm_fit}
\title{Convert \code{\link{crm_fit}} to instance of
\code{\link[coda]{mcmc.list}}}
\usage{
\method{as.mcmc.list}{crm_fit}(crm_fit, ...)
}
\arguments{
\item{crm_fit}{Object of class \code{\link{crm_fit}}}

\item{...}{Extra variables that are passed onwards.}
}
\value{
Object of class \code{\link[coda]{mcmc.list}}
}
\description{
This function allows trialr to use tidybayes functions.
}
