% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augbin_2t_1a_fit.R
\name{as_tibble.augbin_2t_1a_fit}
\alias{as_tibble.augbin_2t_1a_fit}
\title{Cast \code{augbin_2t_1a_fit} object to \code{\link[tibble]{tibble}}.}
\usage{
\method{as_tibble}{augbin_2t_1a_fit}(x, ...)
}
\arguments{
\item{x}{Object of class \code{augbin_2t_1a_fit}.}

\item{...}{Extra args passed onwards.}
}
\value{
Object of class \code{\link[tibble]{tibble}}
}
\description{
Cast \code{augbin_2t_1a_fit} object to \code{\link[tibble]{tibble}}.
}
