% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_finding_paths.R
\name{as_tibble.dose_finding_paths}
\alias{as_tibble.dose_finding_paths}
\title{Cast \code{dose_finding_paths} object to \code{\link[tibble]{tibble}}.}
\usage{
\method{as_tibble}{dose_finding_paths}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dose_finding_paths}.}

\item{...}{Extra args passed onwards.}
}
\value{
Object of class \code{\link[tibble]{tibble}}
}
\description{
Cast \code{dose_finding_paths} object to \code{\link[tibble]{tibble}}.
}
