% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_prob_success.R
\name{binary_prob_success}
\alias{binary_prob_success}
\title{Calculate the binary probability of success.}
\usage{
binary_prob_success(x, ...)
}
\arguments{
\item{x}{an R object of class \code{"augbin_fit"}}

\item{...}{arguments passed to other methods}
}
\description{
Calculate the binary probability of success.
}
