% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{closest_to_target}
\alias{closest_to_target}
\title{Get index of element in vector with value closest to a target}
\usage{
closest_to_target(vector, target)
}
\arguments{
\item{vector}{Identify element in this numeric vector}

\item{target}{numeric target}
}
\value{
an integer indexing \code{vector}
}
\description{
Get index of element in vector with value closest to a target
}
\examples{
closest_to_target(c(0.1, 0.2, 0.3), 0.05)  # 1
closest_to_target(c(0.1, 0.2, 0.3), 0.22)  # 2
closest_to_target(c(0.1, 0.2, 0.3), -0.05) # 1
closest_to_target(c(0.1, 0.2, 0.3), 8) # 3
}
