% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_fit.R
\docType{class}
\name{crm_fit-class}
\alias{crm_fit-class}
\alias{crm_fit}
\title{Class of model fit by \pkg{trialr} using the CRM dose-finding design.}
\usage{
crm_fit(dose_indices, prob_tox, median_prob_tox, prob_mtd,
  recommended_dose, dat, fit, samples = NULL)
}
\arguments{
\item{dose_indices}{A vector of integers representing the dose-levels under
consideration.}

\item{prob_tox}{The posterior mean probabilities of toxicity at doses 1:n;
a vector of numbers between 0 and 1.}

\item{median_prob_tox}{The posterior median probabilities of toxicity at doses
1:n; a vector of numbers between 0 and 1.}

\item{prob_mtd}{The posterior probability that each dose is the MTD, by the
chosen model; a vector of numbers between 0 and 1. This probability reflects
the uncertainty remaining in the parameter distributions, whereas
\code{prob_tox} and \code{median_prob_tox} do not.}

\item{recommended_dose}{An integer representing the dose-level that is
recommended for the next patient or cohort. Contrast to
\code{modal_mtd_candidate}.}

\item{dat}{Object \code{\link{crm_params}} containing data passed to
\code{\link[rstan:sampling]{sampling}}.}

\item{fit}{An object of class \code{\link[rstan:stanfit]{stanfit}},
containing the posterior samples.}

\item{modal_mtd_candidate}{An integer representing the dose-level most likely
to be the MTD, i.e. the dose-level that maximises \code{prob_mtd}.}

\item{entropy}{The value of the entropy implied by prob_mtd. If the values of
prob_mtd are equal, entropy is maximised, taking value log(n) where n is
the number of doses, and we could not be more ignorant about the identity of
the maximum tolerable dose. If prob_mtd is a vector of zeroes except for a
single value of 1, then entropy is minimisied, taking value 0, and we are
sure of the identity of the maximum tolerable dose.}

\item{An}{optional \code{data.frame} like object of samples.}
}
\description{
Class of model fit by \pkg{trialr} using the CRM dose-finding design.
}
\details{
See \code{methods(class = "crm_fit")} for an overview of available
methods.
}
\seealso{
\code{\link{stan_crm}}
}
