% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efftox_parameters_demo.R
\name{efftox_parameters_demo}
\alias{efftox_parameters_demo}
\title{Get parameters to run the EffTox demo}
\usage{
efftox_parameters_demo()
}
\value{
a \code{list} of parameters, described in \code{efftox_params}
}
\description{
Get parameters to run the EffTox demo. These match those used
to demonstrate EffTox in Thall et al. 2014.
}
\examples{
dat <- efftox_parameters_demo()
names(dat)
dat$real_doses == c(1, 2, 4, 6.6, 10)

}
\references{
Thall, Herrick, Nguyen, Venier & Norris. 2014, Effective sample
size for computing prior hyperparameters in Bayesian phase I-II dose-finding
}
\seealso{
\code{\link{efftox_params}}
}
