% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRM.R
\name{gather_samples.crm_fit}
\alias{gather_samples.crm_fit}
\title{Extract tall data.frame of posterior prob_tox samples.}
\usage{
gather_samples.crm_fit(x)
}
\arguments{
\item{x}{\code{\link{crm_fit}} object.}
}
\value{
data.frame
}
\description{
Extract tall data.frame of posterior prob_tox samples.
}
