% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{gather_samples.efftox_fit}
\alias{gather_samples.efftox_fit}
\title{Extract tall data.frame of posterior variable samples.}
\usage{
gather_samples.efftox_fit(x, pars = "utility")
}
\arguments{
\item{x}{\code{\link{efftox_fit}} object.}

\item{pars}{One of 'utility', 'prob_tox', 'prob_eff', 'prob_acc_eff',
or 'prob_acc_tox'}
}
\value{
data.frame
}
\description{
Extract tall data.frame of posterior variable samples.
}
