% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_fit.R
\name{plot.crm_fit}
\alias{plot.crm_fit}
\title{Plot an crm_fit}
\usage{
\method{plot}{crm_fit}(x, pars = "prob_tox", ...)
}
\arguments{
\item{x}{\code{\link{crm_fit}} object to plot.}

\item{pars}{Parameters to plot. Plots utility scores by default.}

\item{...}{Extra parameters, passed onwards.}
}
\value{
A plot
}
\description{
Plot an crm_fit
}
