% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{plot.efftox_fit}
\alias{plot.efftox_fit}
\title{Plot an efftox_fit}
\usage{
\method{plot}{efftox_fit}(x, pars = "utility", ...)
}
\arguments{
\item{x}{\code{\link{efftox_fit}} object to plot.}

\item{pars}{Parameters to plot. Plots utility scores by default.}

\item{...}{Extra parameters, passed onwards.}
}
\value{
A plot
}
\description{
Plot an efftox_fit
}
