% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_fit.R
\name{print.crm_fit}
\alias{print.crm_fit}
\title{Print crm_fit object.}
\usage{
\method{print}{crm_fit}(x, ...)
}
\arguments{
\item{x}{\code{\link{crm_fit}} object to print.}

\item{...}{Extra parameters, passed onwards.}
}
\description{
Print crm_fit object.
}
