% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{print.efftox_fit}
\alias{print.efftox_fit}
\title{Print efftox_fit object.}
\usage{
\method{print}{efftox_fit}(x, ...)
}
\arguments{
\item{x}{\code{\link{efftox_fit}} object to convert.}

\item{...}{Extra parameters, passed onwards.}
}
\description{
Print efftox_fit object.
}
