% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbg_fit.R
\name{print.nbg_fit}
\alias{print.nbg_fit}
\title{Print nbg_fit object.}
\usage{
\method{print}{nbg_fit}(x, ...)
}
\arguments{
\item{x}{\code{nbg_fit} object to print.}

\item{...}{Extra parameters, passed onwards.}
}
\description{
Print nbg_fit object.
}
