% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_success.R
\name{prob_success}
\alias{prob_success}
\title{Calculate the probability of success.}
\usage{
prob_success(fit, ...)
}
\arguments{
\item{fit}{an R object of class \code{"augbin_fit"}}

\item{...}{arguments passed to other methods}
}
\description{
Calculate the probability of success.
}
