\name{Rcpp_model_BebopInPeps2-class}
\alias{Rcpp_model_BebopInPeps2}
\alias{Rcpp_model_BebopInPeps2-class}
\title{Compiled Stan model for BEBOP implementation in PePS2 clinical trial}
\description{
Compiled Stan model for BEBOP implementation in PePS2 clinical trial.
}
\seealso{
\code{\link{model_BebopInPeps2}}
}
