\name{Rcpp_model_ThallHierarchicalBinary-class}
\alias{Rcpp_model_ThallHierarchicalBinary-class}
\alias{Rcpp_model_ThallHierarchicalBinary}
\title{Compiled Stan model for Thall et al.'s hierarchical Bayesian model for binary data}
\description{
Compiled Stan model for Thall et al.'s hierarchical Bayesian model for binary data
}
\seealso{
\code{\link{model_ThallHierachicalBinary}}
}
