% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlkjcorr.R
\name{rlkjcorr}
\alias{rlkjcorr}
\title{Sample LKJ correlation matrices.}
\usage{
rlkjcorr(n, K, eta = 1)
}
\arguments{
\item{n}{Number of matrices to sample.}

\item{K}{dimenstion of matrix to sample.}

\item{eta}{Distribution parameter}
}
\value{
matrix
}
\description{
This function was copied from Richard McElreath's rethinking package hosted
at https://github.com/rmcelreath/rethinking. In turn, he appears to have
copied it from Ben Bolker's rLKJ function from the emdbook package, although
I cannot find it there (else I would have imported it).
}
