% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_augbin_demo.R
\name{stan_augbin_demo}
\alias{stan_augbin_demo}
\title{Simple helper function to demonstrate fitting of an Augmented Binary model.}
\usage{
stan_augbin_demo()
}
\value{
instance of \code{\link{augbin_fit}}
}
\description{
This function exist mostly to demonstrate things you can do to instances of
\code{\link{augbin_fit}} without having to paste into each example the not
inconsiderable blob of code to sample outcomes and fit the model.
}
\examples{
\dontrun{
fit <- stan_augbin_demo()
# I told you it was simple.
}
}
\seealso{
\code{\link{stan_augbin}}
  \code{\link{augbin_fit}}
  \code{\link{prior_predictive_augbin_2t_1a}}
  \code{\link[rstan:sampling]{sampling}}
}
