% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRM.R
\name{summary.crm_fit}
\alias{summary.crm_fit}
\title{Obtain summary of an crm_fit}
\usage{
\method{summary}{crm_fit}(object, ...)
}
\arguments{
\item{...}{Extra parameters, passed onwards.}

\item{x}{\code{\link{crm_fit}} object to summarise.}
}
\value{
A summary object.
}
\description{
Obtain summary of an crm_fit
}
