% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EffTox.R
\name{summary.efftox_fit}
\alias{summary.efftox_fit}
\title{Obtain summary of an efftox_fit}
\usage{
\method{summary}{efftox_fit}(object, ...)
}
\arguments{
\item{...}{Extra parameters, passed onwards.}

\item{x}{\code{\link{efftox_fit}} object to summarise.}
}
\value{
A summary object.
}
\description{
Obtain summary of an efftox_fit
}
