% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thallhierarchicalbinary.R
\name{thallhierarchicalbinary_parameters_demo}
\alias{thallhierarchicalbinary_parameters_demo}
\title{Get parameters to run the demo of Thall Hierarchical Binary model}
\usage{
thallhierarchicalbinary_parameters_demo()
}
\value{
a \code{list} of parameters
}
\description{
Get parameters to run the demo of Thall Hierarchical Binary model.
These match those used to demonstrate EffTox in Thall et al. 2003.
}
\examples{
dat <- thallhierarchicalbinary_parameters_demo()
names(dat)
dat$target_resp == 0.3

}
\references{
Thall et al. 2014, Effective sample size for computing prior
hyperparameters in Bayesian phase I-II dose-finding
}
\seealso{
\code{\link{efftox_params}}
}
