\name{trialr-package}
\alias{trialr-package}
\alias{trialr}
\docType{package}
\title{
\packageTitle{trialr}
}
\description{
\packageDescription{trialr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{trialr}
\packageIndices{trialr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{trialr}

Maintainer: \packageMaintainer{trialr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
