% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_process.R
\name{crm_process}
\alias{crm_process}
\title{Process RStan samples from a CRM model.}
\usage{
crm_process(dat, fit)
}
\arguments{
\item{dat}{An instance of \code{\link{crm_params}}, a list of CRM
parameters.}

\item{fit}{An instance of \code{rstan::stanmodel}, derived by fitting one of
the trialr CRM models.}
}
\value{
An instance of \code{\link{crm_fit}}.
}
\description{
Internal function to process rstan samples from a CRM model to
make inferences about dose-toxicity and which dose should be recommended next.
Typically, this function is not required to be called explicitly by the user
because \code{\link{stan_crm}} will call it implicitly.
}
