% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efftox_analysis_to_df.R
\name{efftox_analysis_to_df}
\alias{efftox_analysis_to_df}
\title{EffTox analysis to data.frame}
\usage{
efftox_analysis_to_df(x)
}
\arguments{
\item{x}{An instance of \code{\link{efftox_fit}}}
}
\value{
a \code{data.frame}
}
\description{
Convenient function to turn an \code{\link{efftox_fit}}
into a \code{data.frame}.
}
\examples{
fit <- stan_efftox_demo(outcome_str = '1N 2E 3B')
df <- efftox_analysis_to_df(fit)
df

}
\seealso{
\code{\link{stan_efftox}}
}
