% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crm_fit.R
\name{summary.crm_fit}
\alias{summary.crm_fit}
\title{Obtain summary of an crm_fit}
\usage{
\method{summary}{crm_fit}(object, ...)
}
\arguments{
\item{object}{\code{\link{crm_fit}} object to summarise.}

\item{...}{Extra parameters, passed onwards.}
}
\value{
A summary object.
}
\description{
Obtain summary of an crm_fit
}
\seealso{
\code{\link{stan_crm}}
}
