% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialr-package.R
\docType{package}
\name{trialr-package}
\alias{trialr-package}
\alias{trialr}
\title{The 'trialr' package.}
\description{
trialr collects in one place Bayesian clinical trial designs and methods.
Models are implemented in Stan and helper functions are provided in R.
}
\references{
Stan Development Team (2018).
RStan: the R interface to Stan. R package version 2.18.2.
https://mc-stan.org/
}
